feat: remove supply page
