--- app/supply/layout.tsx
@@ -1,10 +1,0 @@
-import { RichListProvider } from '@providers/richList';
-import { SupplyProvider } from '@providers/supply';
-
-export default function SupplyLayout({ children }: { children: React.ReactNode }) {
-    return (
-        <SupplyProvider>
-            <RichListProvider>{children}</RichListProvider>
-        </SupplyProvider>
-    );
-}
